fix: Separated virtual page load timing functionality in helper class for PageManager
