+++ package.json
@@ -99,6 +99,7 @@
         "@aws-sdk/signature-v4": "^3.36.0",
         "@aws-sdk/util-hex-encoding": "^3.36.0",
         "@babel/runtime": "^7.16.0",
+        "chromedriver": "^99.0.0",
         "shimmer": "^1.2.1",
         "ua-parser-js": "^0.7.23",
         "uuid": "^8.3.2",
