+++ src/orchestration/__tests__/Orchestration.test.ts
@@ -152,6 +152,8 @@ describe('Orchestration tests', () => {
             pagesToExclude: [],
             pagesToInclude: [],
             recordResourceUrl: true,
+            routeChangeLoadCheckInterval: 100,
+            routeChangeTimeout: 1000,
             sessionEventLimit: 200,
             sessionLengthSeconds: 1800,
             sessionSampleRate: 1,
